Feat: refactor test-utils, add helpers funciton and add new product ced test
